add csrf_token protection

+++ togger/calendar/calendar_api.py
@@ -1,10 +1,12 @@
 import flask_login
 from flask import Blueprint, request, url_for, flash
+from flask_wtf import FlaskForm
 from werkzeug.utils import redirect
 
 from togger.auth import auth_dao
 from togger.auth.models import Role
 from togger.calendar import calendar_dao
+from togger.calendar.calendar_forms import DeleteForm, NewForm, SettingsForm
 
 bp = Blueprint("calendar_api", __name__, url_prefix="/api/v1/calendars")
 
@@ -13,7 +15,8 @@ bp = Blueprint("calendar_api", __name__, url_prefix="/api/v1/calendars")
 @flask_login.login_required
 def post_calendars():
     if '_method' not in request.form:
-        calendar_dao.create(request.form['calendarName'])
+        if NewForm().validate_on_submit():
+            calendar_dao.create(request.form['calendarName'])
         return redirect(url_for('main'))
     elif request.form['_method'] == "DELETE":
         return delete_calendar()
@@ -23,7 +26,8 @@ def post_calendars():
 @flask_login.login_required
 @auth_dao.has_role(Role.OWNER)
 def delete_calendar():
-    calendar_dao.delete()
+    if DeleteForm().validate_on_submit():
+        calendar_dao.delete()
     return redirect(url_for('main'))
 
 
@@ -44,20 +48,22 @@ def post_share():
 @flask_login.login_required
 @auth_dao.has_role(Role.OWNER)
 def change_share():
-    user_id = request.form['userId']
-    role_name = request.form['roleNameShares']
-    calendar_dao.change_share(user_id, int(role_name))
+    if FlaskForm().validate_on_submit():
+        user_id = request.form['userId']
+        role_name = request.form['roleNameShares']
+        calendar_dao.change_share(user_id, int(role_name))
     return '', 204
 
 
 @bp.route('/settings', methods=['POST'])
 @flask_login.login_required
 def post_settings():
-    settings = {'scrollTime': request.form['scrollTime'], 'firstDay': request.form['firstDay'],
-                'slotMinTime': request.form['slotMinTime'], 'slotMaxTime': request.form['slotMaxTime'],
-                'nextDayThreshold': request.form['nextDayThreshold']}
-    if calendar_dao.save_settings(settings):
-        flash("Settings saved", 'success')
+    if SettingsForm().validate_on_submit():
+        settings = {'scrollTime': request.form['scrollTime'], 'firstDay': request.form['firstDay'],
+                    'slotMinTime': request.form['slotMinTime'], 'slotMaxTime': request.form['slotMaxTime'],
+                    'nextDayThreshold': request.form['nextDayThreshold']}
+        if calendar_dao.save_settings(settings):
+            flash("Settings saved", 'success')
     return redirect(url_for('render_settings'))
